old system/requirements.txt
# Policy Analysis Agent Requirements
# Python >= 3.11

# A2A SDK and Server
a2a-sdk[http-server]==0.3.10

# Web Frameworks
fastapi==0.120.4
starlette==0.49.3
uvicorn==0.38.0
sse-starlette==3.0.3
aiofiles==25.1.0

# LangChain + LangGraph stack
langchain-core==1.0.2
langchain-openai==1.0.1
langchain-community==0.4.1
langgraph==1.0.2

# Data and Config
pydantic==2.12.3
pydantic-settings==2.11.0
python-dotenv==1.2.1
click==8.3.0
anyio==4.11.0
httpx==0.28.1

# Database
sqlalchemy==2.0.44

# PDF and Image Processing
pypdf==6.1.3
pdf2image==1.17.0
pillow==12.0.0

# Frontend (Streamlit)
streamlit==1.51.0

# System Monitoring
psutil==6.1.0
